Rename RenderCommand.h to RenderCommandQueue.h
